refactor 바우처 인터페이스 제거
바우처 인터페이스를 제거 후 추상클래스에 추상메서드 추가
