+++ src/test/java/com/dev/bootbasic/voucher/domain/PercentDiscountVoucherTest.java
@@ -19,10 +19,10 @@ class PercentDiscountVoucherTest {
     void validateDiscountAmountSuccessTest(int discountAmount) {
         UUID id = UUID.randomUUID();
 
-        AbstractVoucher voucher = PercentDiscountVoucher.of(id, discountAmount);
+        Voucher voucher = PercentDiscountVoucher.of(id, discountAmount);
 
         assertThat(voucher)
-                .extracting(AbstractVoucher::getId, AbstractVoucher::getDiscountAmount)
+                .extracting(Voucher::getId, Voucher::getDiscountAmount)
                 .containsExactly(id, discountAmount);
     }
